pyproject.toml
[project]
name = "multi-ai-agent-gcp"
version = "0.1.0"
description = "Multi-AI Agent using Groq LLM and Tavily Search with FastAPI backend and Streamlit frontend - Deployable on GCP Cloud Run"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.6",
    "langchain>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-groq>=1.1.2",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.8",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "uvicorn>=0.40.0",
]
